Se agregan las excepciones

+++ src/main/java/com/proyecto/personal/repositorio/EmpleadosRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.personal.model.Empleados;
 
 @Repository
-public interface EmpleadosRepositorio extends JpaRepository <Empleados, int>{
+public interface EmpleadosRepositorio extends JpaRepository<Empleados, Long>{
 
 }
